0.2.0
Created mutithreaded webcrawler that remembers the last URLs and is able
to resume, and is able to create appropriate number of threads. All of
these requirements were tested using unit tests in the BotTest.java file
overall the Bot class is the implementation of the webcrawler  and
contains the logic for crawling web pages while Main class file is
designed to test the webcrawler. comments are provided for further
explanation of the methods used.